move graphics drawing commands to MainGenerator, add extensions for writing text, log lines write all colors
